Fixed a bug where the timeline wouldn't stretch to completely 100%
I think it was because of a rounding issue. To solve, we basically just
take the slice that was missing after the rounding error happened, and
divide it up equally among the others to redistribute the width evenly.
I feel dirty for doing this. I would like to actually just solve the
rounding issue somehow.