Move parsers to their runs' module file
This is still not perfect but it's a step in that direction I believe. I think that it's clean to be able to
compartmentalize all code required for any given program to one file (given that it's not too big), and I think that
having that one file be the only addition necessary to add support for a new program (plus a single-word addition in
run.rb) lowers the barrier to entry for others adding support.

This commit accomplishes that. But I don't think that app/models is the place for that file, and I don't like that we
have that file contain both <program>::Run (a model) and <program>::Parser (a tool). I'm not sure what the final
solution will look like.