Fix weird links for games with periods in them
Closes #233.

Rails treats periods special, because it wants to infer response format from anything after it. So it eats that stuff up
and only gives us the stuff before the period. Previously we said ok to this and truncated period-links we produced
so that Rails wouldn't eat up something weird. But now we will just disable that Rails functionality so we can have
better links.