New color method and graph detection
The graph code will run on any page now that has the appropriate div
ID's located on them.  Colors are now passed through a gon variable that
is originally defined in RunsHelper.  This allows the changing of all
colors from one location.  It parses the CSS styles for the matching
class and pulls the background color from that.