Try out explicitly fetching from DynamoDB
Not sure how I'm going to feel about this pattern. I want to move more
fetching logic over to DynamoDB and I also want it to be obvious where
you're fetching data from when you fetch it (like in views), so that we
don't accidentally do large parses without realizing it. But not sure if
this is the right way to accomplish that. We'll try it out for a bit and
see.

If we go with this we'll obv eventually move the DynamoDB fetch code out
of `parse`.