Allow claiming anonymous runs
Allows users to self-claim runs they uploaded while logged out, which is
the most common support request I get from people that I have to fulfill
manually.

Somehow I had the foresight 2 years ago to add [a single line of
code][1] that records claim tokens in localStorage when a run is
uploaded, but simultaneously the laziness to waste two years before
putting it to any use. So theoretically you can use this feature to
claim runs uploaded anonymously ever since.

Would be nice to also have a page that lists claimable runs, but that's
more complex.

[1]: https://github.com/glacials/splits-io/commit/191af4635d3d71f4f6cfe8e198ac67e3db676e99#diff-471b2f654e508c40fc08adf4ea19ea00R13