Change how to fill in Twitch credentials again
Looks like the `--skip-worktree` has a downside where if you change the
file, `--skip-worktree` it, then try to checkout another branch, Git
may still complain at you that you have unstashed changes. (And stashing
doesn't even include the Dockerfile if it's been `--skip-worktree`d.)

I really like the idea of not needing to do extra stuff to get splits-io
running the first time, so I'm going to try out this method where the
extra work is still in the getting-login-working step, but is otherwise
more like what we used to do.