Make search in navbar expand to fill area
The shorter search bar kind of bugged me, especially since it expanded
to full-width in the mobile overflow menu. This uses flexbox to let the
search bar expand to fill the entire area at any window width greater
than or equal to 768px.

Full disclosure: this was NOT tested very well. I could not get Docker
to start locally (on Windows) and thus was only able to test this via
the ol' browser tools. I don't even know if this was the correct
stylesheet to edit.