Fix sql depreciations introduced in 5.2
Misattributed to pg_search, when previously looked at it was checking the wrong line.

In docker-compose this fixes running the webpack server for my VM.  For some reason with `bundle exec` it throws cannot find ruby.  It effectively does the same thing since the binstub loads everything that it needs