Redo upload page design, use site_title everywhere
Redesigns the upload page to use Bootstrap cards, and merges the convert
page into it.

Introduces `site_title` method in ApplicationHelper, and replaces all
breadcrumbs on the whole site to use it instead of manually writing
"Splits I/O". This has the benefit of making it more obvious when you're
on Beta or local dev because it will say "Splits I/O (Local)" etc.

Made breadcrumbs consistently formatted while I was at it, and ditched
the `title` method in ApplicationHelper because it didn't save many
keystrokes and adds another layer of non-Rails-default things to learn
when developing on splits-io.