Bring segment stats to the run page
Continuing to move stuff off the back-alley stats page and bring it to
the limelight of the run page. This branch moves segment stats --
standard deviation, mean, median, nth percentiles, etc. -- to the run
page. They are accessed in the same dropdown as the segment duration
charts.

This also adds a new feature that I slipped into master on the standard
stats page before this, which is telling lighting up areas of
improvement.