Exchange format
Adds support for the Splits I/O Exchange Format, first detailed in #337.
Importing is supported by livesplit-core 0.10, exporting is using an ERB
file (with plans to move that to an ActiveModel Serializer once things
settle).

livesplit-core 0.10 also added support for worstrun and Source Live
Timer, so I added modules for those as well.