Become our own OAuth client
Makes the frontend an OAuth consumer of the backend. This gives it its
own access token it can use to interact with the API, instead of
shortcutting it by allowing the API to accept cookies under special
circumstances.

This is a bit rough right now, as the frontend's access token needs to
be refreshed once in a while. This refresh is done by sending the user
through the OAuth flow. We can auto-accept the grant since we're a
special client, but there's still a short flash of the in-between page.

We can clean this up incrementally later by doing things like using
`Turbolinks.visit` instead of `window.location =` (although I couldn't
immediately get it to work here), getting this JavaScript to execute
super early on in pageload, or perhaps using an iframe instead of the
whole browser.