Change comparison patron requirements
Change comparisons from being freely available for self-self comparisons
and patron-only for all other comparisons to allowing timeline-only
comparisons for all and patron-only to view the rest of the comparison
tools.

This preserves the freeness of the current comparison page, which is a
timeline comparison only.

Also refactored a lot of the page to allow for any number of
simultaneous comparisons in the future.