Add Stripe integration
Adds Stripe integration for $4 and $6 tiers, and adds a landing page to
show them along with special callouts for the coaching tier and desires
to work with organizations.

Also adds a privacy policy, more consistent callouts for subscribing
when you reach a place that would've had a feature, and a better
logged-in new-user frontpage.