Fix various API bugs
Fixed various API bugs and documentation inconsistencies. (Thanks
@cryze92!)

- Add `GET /api/v4/runner` to retrieve the currently authenticated runner
- Move documentation notes on each model's ID type (e.g. UUID, base 36 string) to within that model's section
- Fix docs on `/api/v4/races/:race/entry` -> `/api/v4/races/:race/entries/:entry`
- Fix JSON schema enums being specified wrong
- Fix some missing timers in test files
- Add text for the OAuth grant page for the `manage_race` scope
- Allow `twitch_id`, `twitch_name` to be null now
- Fix race visibility being a number in docs

I was not able to find the bug where an entry's `created_at` is null.

Also for now I'm leaving out the bug where a user with the race page
open who changes their entry from the API, causes the website to not
know that the change was made, as that's a bit more involved a fix.