Fix some auto-highlights never triggering
Because I missed a zero_to_nil call when refactoring absent times to be
nil (they were zero before), if the run was realtime-only the
HighlightSuggestion was looking for a RunHistory with a gametime of nil
(not finding the RunHistory because it was 0), and same for if the run
was gametime-only.

So basically, only runs with both realtime and gametime were getting
auto-highlight suggestions since that refactor.