Update upload UX (#620)
* Update upload UX

Updates the visuals of uploading as well as offloads the parsing into an
asynchronous job. Allows the run page to be completely loaded before
parsing has finished, and shows a spinner until it completes, notified
via Action Cable.

This sets us up to take parsing out of the web application and into a
separate, independably scalable service like Lambda or a worker box.
That world will fix the issue of a user uploading a run so complex that
it crashes or temporarily disables a/the web box. This branch doesn't
get us completely there, but should help nonetheless as the worker is in
its own Docker container.

* CR comments

* CR comments