Pricing overhaul, race splitting
* Add splitting to races

Adds a Split button to races. This button will split if there is at
least one segment left in the run, or create a segment then split if
not. It will also create a run if one isn't yet attached to the race
entry. These two features together can be used to build splits adhoc
during a blind or semi-blind races.

The API allows you to choose whether to enable this smart behavior by
passing `more=1`. When not included, the API will only split in the
classical sense, completing the run if you're on the last segment.

Includes several other race improvements:

- Adds timelines to races for each runner splitting
- Timelines (both in & out of races) now have smarter segment widths for
  in-progress runs (they were previously all equivalent widths)
- Adds a "Syncing" spinner to the race page that shows whenever you've
  performed an action but it hasn't reflected on the page yet due to
  the worker and/or ActionCable taking an extra bit
- The race page is now more aggressive about hiding itself (and showing
  a spinner) while loading / while navigating away from the page, rather
  than showing outdated information cached by Turbolinks
- The Create Race button no longer shows an error symbol after being
  clicked & transitioning to the new race page, even though there was no
  error
- Can no longer start editing a race after it has started

* CR comments

* Rubocop

* Pricing overhaul

* Clean up an unused ad file, fix some copy

* Fix highlight button showing for non-owners

* Fix inconsistent button styling on game page

* CR comments

* Fix split button

* Hide Pricing link if already subscribed

* Fix advanced charts

* Fix a logged-out 500

* Fix subscriptions section on Settings page

* Fix another logged-out 500