Tweak design to be more mobile-friendly
Tweaks a lot of the site design to be more friendly when viewing on
mobile. On the whole, the desktop experience should be unchanged except
for a couple of sidegrades like the run table.

There's still some more I want to do, including probably replicate what
I did to the run table to all other similar tables, but this is a good
cut-off point for this branch both in terms of reviewability and ability
to go to production on its own before it gets too big.