Add affiliate links
Adds an affiliate links feature to support upcoming partners. Navigate
to splits.io/test, get redirected to the subscriptions page with a
special code. If you sign up, the referring partner (test) will get a
kickback.

This works by applying a Stripe Coupon, which will be tied 1-to-1 to a
partner. We can later query against coupons to determine payouts.

Stripe Checkout, the pre-baked hosted checkout flow we use, apparently
doesn't yet support Coupons. So we have to go through the Checkout flow
without it then apply the Coupon afterwards.

This is decent if the Coupon isn't actually applying a meaningful
discount (right now it's the minimum, $0.01), but if we long-term want
to give users discounts as an incentive to use affiliate links, they'll
still see full price in the Checkout flow (and be charged full price for
the first month) so we'd probably want to find a better solution. Mostly
hoping Stripe Checkout implements Coupon support before then.