Make game/category selectors Vue components
Migrates the three areas on the site where you can autocomplete a
game/category -- search, edit run, and edit race -- to Vue components
with a decent amount of reusability.

This should incidentally fix the bug where the search bar gets
ludicrously tall, since we're not powering it with that old library
anymore.