Fiddle with YouTube button design & tooltip
Moves button down to card footer and make it save the video as well,
since the first thing people will do after clicking it is probably leave
the page and forget to save it.

Adds a tooltip to the offset field, an example to the button tooltip,
and changes button text to "Copied!" when timestamps are copied.