Don't try to cancel non-active subscriptions
Well that was easy! Fixes #715. We were 500ing on the Stripe webhook
because Stripe gave us errors when we tried to cancel subscriptions that
had already ended. They had not been "canceled" (which is an explicit
thing the user does), but _had_ been "ended" because of card declines,
so a cancellation by us didn't make sense.