Allow trials without credit cards
Currently trial subscriptions are handled completely by Stripe, and they
don't have an option to allow them without entering payment information
first.

This branch moves trial subscriptions to be handled by us, so that we
don't need to have that requirement in place. It also refactors the code
that decides which users have which features, because it was getting
pretty muddy having so many features with so many possible grantors of
that feature.